[Playlists] Split RAM playlist handling into own files
